[clean].env.testingの修正とテストの階層を修正
