+++ .env.testing
@@ -9,11 +9,11 @@ LOG_DEPRECATIONS_CHANNEL=null
 LOG_LEVEL=debug
 
 DB_CONNECTION=pgsql
-DB_HOST=127.0.0.1
+DB_HOST=db-test
 DB_PORT=5432
-DB_DATABASE=test_db
-DB_USERNAME=test_user
-DB_PASSWORD=test_password
+DB_DATABASE=testing
+DB_USERNAME=test
+DB_PASSWORD=password
 
 BROADCAST_DRIVER=log
 CACHE_DRIVER=file
