+++ tests/Feature/api/Article/StoreTest.php
@@ -6,9 +6,11 @@ use App\Models\Article;
 use App\Models\Tag;
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
+use Symfony\Component\HttpFoundation\Response as ResponseAlias;
 use Tests\TestCase;
+use Illuminate\Support\Facades\Storage;
 
-class ArticleControllerUpdateTest extends TestCase
+class StoreTest extends TestCase
 {
     use RefreshDatabase;
 
@@ -18,6 +20,8 @@ class ArticleControllerUpdateTest extends TestCase
     protected function setUp(): void
     {
         parent::setUp();
+
+        Storage::fake('s3');
 
         Tag::factory()
             ->count(20)
@@ -55,35 +59,25 @@ class ArticleControllerUpdateTest extends TestCase
      *
      * @return void
      */
-    public function 記事編集ができること(): void
+    public function 記事投稿ができること(): void
     {
         $tags = Tag::take(3)->get();
-        // 記事を投稿
-        $article = Article::factory()->create([
-            'title' => 'タイトル',
-            'content' => '本文',
-            'user_id' => $this->user->id,
-        ]);
-        // 編集前のDBの状態を確認
-        $this->assertDatabaseHas('articles', [
-            'id' => $article->id,
-            'title' => 'タイトル',
-            'content' => '本文',
-            'user_id' => $this->user->id,
+
+        // actingAsでログイン状態にする
+        $this->actingAs($this->user);
+
+        $response = $this->postJson('/api/articles', [
+            'title' => 'test title',
+            'content' => 'test content',
+            'tags' => [
+                $tags[0]->id,
+                $tags[1]->id,
+                $tags[2]->id,
+            ],
         ]);
 
-        $response = $this->actingAs($this->user)
-            ->putJson('/api/articles/'. $article->id, [
-                'title' => 'タイトル編集',
-                'content' => '本文編集',
-                'tags' => [
-                    $tags[0]->id,
-                    $tags[1]->id,
-                ],
-            ]);
-        // 200レスポンスが返ってくること
-        $response->assertStatus(200)
-        // レスポンスの構造を確認
+        $response->assertStatus(201)
+            // レスポンスの構造を確認
             ->assertJsonStructure([
                 'data' => [
                     'title',
@@ -98,36 +92,38 @@ class ArticleControllerUpdateTest extends TestCase
                     ],
                 ],
             ])
-        // レスポンスの内容を確認
+            // レスポンスの内容を確認
             ->assertJson([
                 'data' => [
-                    'title' => 'タイトル編集',
-                    'created_at' => $article->created_at,
+                    'title' => 'test title',
                     'display_name' => $this->user->display_name,
                     'icon_path' => $this->user->icon_path,
                     'number_of_likes' => 0,
                     'tags' => [
                         ['name' => $tags[0]->name],
                         ['name' => $tags[1]->name],
+                        ['name' => $tags[2]->name],
                     ],
                 ],
             ]);
-        // 編集後DBに保存されていることを確認
+        // DBに保存されていることを確認
         $this->assertDatabaseHas('articles', [
-            'id' => $article->id,
-            'title' => 'タイトル編集',
-            'content' => '本文編集',
+            'title' => 'test title',
+            'content' => 'test content',
             'user_id' => $this->user->id,
         ]);
         // 中間テーブルに保存されていることを確認
         $this->assertDatabaseHas('article_tag', [
-            'article_id' => $article->id,
+            'article_id' => Article::first()->id,
             'tag_id' => $tags[0]->id,
         ]);
-        // 中間テーブルに保存されていることを確認
         $this->assertDatabaseHas('article_tag', [
-            'article_id' => $article->id,
+            'article_id' => Article::first()->id,
             'tag_id' => $tags[1]->id,
+        ]);
+        $this->assertDatabaseHas('article_tag', [
+            'article_id' => Article::first()->id,
+            'tag_id' => $tags[2]->id,
         ]);
     }
 
@@ -144,33 +140,16 @@ class ArticleControllerUpdateTest extends TestCase
         array $data,
         array $expectedErrors
     ): void {
-        // 記事を投稿
-        $article = Article::factory()->create([
-            'title' => 'タイトル',
-            'content' => '本文',
-            'user_id' => $this->user->id,
-        ]);
-        // 編集前のDBの状態を確認
-        $this->assertDatabaseHas('articles', [
-            'id' => $article->id,
-            'title' => 'タイトル',
-            'content' => '本文',
-            'user_id' => $this->user->id,
-        ]);
-        // ログインして記事を編集
-        $response = $this->actingAs($this->user)
-            ->putJson('/api/articles/'. $article->id, $data);
-        // 422レスポンスが返ってくること
+        $this->actingAs($this->user);
+        $response = $this->postJson('/api/articles', $data);
+
         $response->assertStatus(422)
-        // エラーメッセージを確認
+
+            // エラーメッセージの内容を確認
             ->assertJsonValidationErrors($expectedErrors);
-        // // DBに編集が反映されていないことを確認
-        $this->assertDatabaseHas('articles', [
-            'id' => $article->id,
-            'title' => 'タイトル',
-            'content' => '本文',
-            'user_id' => $this->user->id,
-        ]);
+
+        // DBに保存されていないことを確認
+        $this->assertDatabaseCount('articles', 0);
     }
 
     /**
@@ -281,125 +260,24 @@ class ArticleControllerUpdateTest extends TestCase
      *
      * @return void
      */
-    public function 未ログインでは記事編集ができないこと(): void
+    public function 未ログインでは記事投稿ができないこと(): void
     {
-        $tags = Tag::take(3)->get();
-        // 記事を投稿
-        $article = Article::factory()->create([
-            'title' => 'タイトル',
-            'content' => '本文',
-            'user_id' => $this->user->id,
-        ]);
-        // 編集前のDBの状態を確認
-        $this->assertDatabaseHas('articles', [
-            'id' => $article->id,
-            'title' => 'タイトル',
-            'content' => '本文',
-            'user_id' => $this->user->id,
-        ]);
-        // 未ログインで記事を編集
-        $response = $this->putJson('/api/articles/'. $article->id, [
-            'title' => 'タイトル編集',
-            'content' => '本文編集',
+        $tags = Tag::all();
+
+        $response = $this->postJson('/api/articles', [
+            'title' => 'test title',
+            'content' => 'test content',
             'tags' => [
                 $tags[0]->id,
                 $tags[1]->id,
+                $tags[2]->id,
             ],
         ]);
-        // 401レスポンスが返ってくること
-        $response->assertStatus(401);
-        // DBに編集が反映されていないことを確認
-        $this->assertDatabaseHas('articles', [
-            'id' => $article->id,
-            'title' => 'タイトル',
-            'content' => '本文',
-            'user_id' => $this->user->id,
-        ]);
-    }
 
-    /**
-     * @test
-     *
-     * @return void
-     */
-    public function 他のユーザーの記事を編集しようとした場合403エラーが返り編集できないこと(): void
-    {
-        $tags = Tag::take(3)->get();
-        // 記事を投稿
-        $article = Article::factory()->create([
-            'title' => 'タイトル',
-            'content' => '本文',
-            'user_id' => $this->user->id,
-        ]);
-        // 編集前のDBの状態を確認
-        $this->assertDatabaseHas('articles', [
-            'id' => $article->id,
-            'title' => 'タイトル',
-            'content' => '本文',
-            'user_id' => $this->user->id,
-        ]);
-        // 別のユーザーを作成
-        $anotherUser = User::factory()->create();
-        // 別のユーザーでログイン
-        $response = $this->actingAs($anotherUser)
-            ->putJson('/api/articles/'. $article->id, [
-                'title' => 'タイトル編集',
-                'content' => '本文編集',
-                'tags' => [
-                    $tags[0]->id,
-                    $tags[1]->id,
-                ],
-            ]);
-        // 403レスポンスが返ってくること
-        $response->assertStatus(403);
-        // DBに編集が反映されていないことを確認
-        $this->assertDatabaseHas('articles', [
-            'id' => $article->id,
-            'title' => 'タイトル',
-            'content' => '本文',
-            'user_id' => $this->user->id,
-        ]);
-    }
+        $response->assertStatus(401)
+            ->assertStatus(ResponseAlias::HTTP_UNAUTHORIZED);
 
-    /**
-     * @test
-     *
-     * @return void
-     */
-    public function 存在しない記事_i_dを指定した場合404エラーが返ること(): void
-    {
-        $tags = Tag::take(3)->get();
-        // 記事を投稿
-        $article = Article::factory()->create([
-            'title' => 'タイトル',
-            'content' => '本文',
-            'user_id' => $this->user->id,
-        ]);
-        // 編集前のDBの状態を確認
-        $this->assertDatabaseHas('articles', [
-            'id' => $article->id,
-            'title' => 'タイトル',
-            'content' => '本文',
-            'user_id' => $this->user->id,
-        ]);
-        // ログインして存在しない記事IDを指定して記事を編集
-        $response = $this->actingAs($this->user)
-            ->putJson('/api/articles/'. ($article->id + 1), [
-                'title' => 'タイトル編集',
-                'content' => '本文編集',
-                'tags' => [
-                    $tags[0]->id,
-                    $tags[1]->id,
-                ],
-            ]);
-        // 404レスポンスが返ってくること
-        $response->assertStatus(404);
-        // DBに編集が反映されていないことを確認
-        $this->assertDatabaseHas('articles', [
-            'id' => $article->id,
-            'title' => 'タイトル',
-            'content' => '本文',
-            'user_id' => $this->user->id,
-        ]);
+        // DBに保存されていないことを確認
+        $this->assertDatabaseCount('articles', 0);
     }
 }
